% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azmpdata-package.R
\docType{package}
\name{azmpdata-package}
\alias{azmpdata-package}
\title{azmpdata: A data package to share derived metrics from AZMP research documents}
\description{
A package which allows users to easily access AZMP data. The goal is to allow AZMP research metrics to be easily shared and incorporated into other research.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://casaultb.github.io/azmpdata}
  \item \url{https://github.com/casaultb/azmpdata}
  \item Report bugs at \url{https://github.com/casaultb/azmpdata/issues}
}

}
\author{
\strong{Maintainer}: Benoit Casault \email{Benoit.Casualt@dfo-mpo.gc.ca}

Authors:
\itemize{
  \item Emily Chisholm \email{Emily.Chisholm@dfo-mpo.gc.ca} (\href{https://orcid.org/0000-0002-1072-5879}{ORCID})
  \item Chantelle Layton \email{Chantelle.Layton@dfo-mpo.gc.ca} (\href{https://orcid.org/0000-0002-3199-5763}{ORCID})
  \item Mike McMahon \email{Mike.McMahon@dfo-mpo.gc.ca}
}

Other contributors:
\itemize{
  \item Catherine Johnson \email{Catherine.Johnson@dfo-mpo.gc.ca} [contributor]
}

}
\keyword{internal}
