% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azmpdata.R
\docType{package}
\name{azmpdata}
\alias{azmpdata}
\alias{azmpdata-package}
\title{azmpdata: A package providing data products derived from data collected as part of
The Atlantic Zone Monitoring Program (AZMP).}
\description{
The azmp package provides three categories of data products:
physical, biological and chemical variables.
}
\section{Physical variables}{

}

\section{Biological variables}{

}

\section{Chemical variables}{

}

