% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azmpmap.R
\name{azmpmap}
\alias{azmpmap}
\title{azmpmap}
\usage{
azmpmap()
}
\description{
This generates a leaflet plot of the geospatial files stored at
ftp://ftp.dfo-mpo.gc.ca/AZMP_Maritimes/azmpdata/data/lookup/polygons/.
}
\examples{
\dontrun{
map <- azmpmap()
}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
