% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chlorophyll_inventory_annual_means_hl2}
\alias{chlorophyll_inventory_annual_means_hl2}
\title{Annual means of chlorophyll inventory at Halifax-2 station.}
\format{A data frame with two variables: \code{year}, and \code{chl_0_100}.
The chlorophyll inventory has units of mg chl/m2.}
\usage{
chlorophyll_inventory_annual_means_hl2
}
\description{
The chlorophyll inventory consists in the integrated chlorophyll concentration
 as obtained using trapezoidal integration over the 0-100m layer for individual
 profiles. The annual means are obtained by fitting a linear model of the form
 \code{density ~ year + month} where density is \code{chl_0_100}, and
 \code{year} and \code{month} are factors.
}
\keyword{datasets}
