% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chlorophyll_inventory_timeseries_hl2}
\alias{chlorophyll_inventory_timeseries_hl2}
\title{Timeseries of chlorophyll inventory at Halifax-2 station.}
\format{
A data frame with four variables: \code{year}, \code{month},
  \code{day}, and \code{chl_0_100}. The chlorophyll inventory has
   units of mg chl/m2.
}
\usage{
chlorophyll_inventory_timeseries_hl2
}
\description{
The chlorophyll inventory consists in the integrated chlorophyll concentration
 as obtained using trapezoidal integration over the 0-100m layer for individual
 profiles.
}
\keyword{datasets}
