% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_all_data.R
\name{combine_all_data}
\alias{combine_all_data}
\title{Combine RDA files into package data tables}
\usage{
combine_all_data()
}
\description{
Step 3 of updating `azmpdata`, combines all `.rda` files from data-raw into organized package data tables based on temporal, spatial and categorical scales.
}
\examples{
combine_all_data()

}
\author{
Emily Chisholm
}
