% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Derived_Annual_Broadscale}
\alias{Derived_Annual_Broadscale}
\title{Derived monthly broadscale data}
\format{
An object of class \code{data.frame} with 856 rows and 6 columns.
}
\usage{
Derived_Annual_Broadscale
}
\description{
_Metadata_
* area
* year
* month
}
\details{
_Data_
* sea_surface_temperature_from_satellite
* sea_ice_area

Data variables are described in more detail in variable look-up table (see \code{\link{lookup_variable}})
Derived annual broadscale data

_Metadata_
* area
* year

_Data_
* sea_temperature_at_sea_floor_july
* sea_surface_temperature
* north_atlantic_oscillation
* atlantic_multidecadal_oscillation
* river_flux
* cold_intermediate_layer_volume
* sea_ice_volume
* density_gradient_0_50
* temperature_at_sea_floor
* minimum_temperature_in_cold_intermediate_layer
* salinity
* temperature
* sea_surface_temperature_from_satellite
* sea_surface_temperature_warming
* final_day_of_sea_ice
* sea_ice_area
* start_of_phytoplankton_bloom
* duration_of_phytoplankton_bloom
* amplitude_of_phytoplankton_bloom
* magnitude_of_phytoplankton_bloom
* nitrate_at_sea_Floor
* primary_production_from_satellite
* oxygen_at_sea_floor
* surface_chlorphyll_from_satellite
}
\keyword{datasets}
