% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Derived_Annual_Carbonate}
\alias{Derived_Annual_Carbonate}
\title{Derived Annual Carbonate Chemistry data}
\format{
A dataframe

\describe{
   \item{year}{The year in which data was collected}
   \item{station_name}{The name of the station where data was collected}
   \item{depth_m}{The depth at which data was collected (metres)}
   \item{mean_TA_umolkg}{annual mean per section_name and depth_m of total alkalinity in umol/kg}
   \item{mean_DIC_umolkg}{annual mean per section_name and depth_m of dissolved inorganic carbon in umol/kg}
   \item{mean_pH_total}{annual mean per section_name and depth_m of pH on a total scale}
   \item{mean_pCO2_uatm}{annual mean per section_name and depth_m of pCO2 in uatm}
   \item{mean_substrate_inhibitor_ratio}{annual mean per section_name and depth_m of substrate-inhibitor ratio (mol/umol)}
   \item{mean_omega_aragonite}{annual mean per section_name and depth_m of the saturation state of omegaAragonite (unitless)}
   \item{mean_omega_calcite}{annual mean per section_name and depth_m of the saturation state of omega calcity (untiless)}
   \item{mean_carbonate_system_vulnerability_index}{annual mean per section_name and depth_m of Carbonate System Vulnerability Index (TA − DIC) (µmol/kg).}
}
}
\usage{
Derived_Annual_Carbonate
}
\description{
Annual composites of derived carbonate chemistry data collected at fixed stations by year.
}
\details{
Data collection and analysis was led by Kumiko Azetsu-Scott and data
calculations were done by Martine Lizotte. Annual means are LS-means by year from linear models.
If all values of a variable are positive in a section, a log10 model is used and back transformed.
If any value is non-positive a linear model is used on the original scale.
season and station_name are included in the model when they have >1 levels in that section.
All missing values are replaced by -999.

More details on calculation methodology can be found in
Lizotte, M., Blais, M., Chassé, J., Galbraith, P. S., Hébert, A.-J., Starr, M. 2026. Acidification and CO2-Driven Conditions in the Estuary and Gulf of St. Lawrence During 2024. Can. Tech. Rep. Hydrogr. Ocean. Sci. 0000 : v + x p.
(in publication)
}
\keyword{datasets}
