% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Derived_Annual_Sections}
\alias{Derived_Annual_Sections}
\title{Derived Annual Section data}
\format{
A dataframe
\describe{
    \item{section}{Section name where data was collected}
    \item{year}{The year in which data was collected}

    \item{density}{Discrete density measurements over a range of depths}
    \item{sea_temperature}{Discrete sea temperature measurements over a range of depths}
    \item{salinity}{Discrete salinity measurements over a range of depths}
    \item{integrated_nitrate_0_50}{Nitrate concentrations integrated between 0 and 50 metres}
    \item{integrated_nitrate_50_150}{Nitrate concentrations integrated between 50 and 150 metres}
    \item{integrated_phosphate_0_50}{Phosphate concentrations integrated between 0 and 50 metres}
    \item{integrated_phosphate_50_150}{Phosphate concentrations integrated between 50 and 150 metres}
    \item{integrated_silicate_0_50}{Silicate concentrations integrated between 0 and 50 metres}
    \item{integrated_silicate_50_150}{Silicate concentrations integrated between 50 and 150 metres}
    \item{integrated_chlorphyll_0_100}{Chlorophyll concentrations integrated between 0 and 100 metres}
}
}
\usage{
Derived_Annual_Sections
}
\description{
Derived data collected along AZMP transects each year.
}
\details{
The variables are as follows:

The data can be cited as follows:
  Casault, B., Johnson, C., Devred, E., Head, E., Cogswell, A., and
  Spry, J. 2020. Optical, Chemical, and Biological Oceanographic Conditions
  on the Scotian Shelf and in the Eastern Gulf of Maine during 2018. DFO Can.
  Sci. Advis. Sec. Res. Doc. 2020/037. v + 66 p.

  Hebert, D., Pettipas, R., and Brickman, D. 2020. Physical Oceanographic
  Conditions on the Scotian Shelf and in the Gulf of Maine during 2018. DFO
  Can. Sci. Advis. Sec. Res. Doc. 2020/036 iv + 52 p.
}
\note{
\describe{
   \item{time_scale}{Annual}
   \item{regional_scale}{Section}
   \item{category}{ biological, biochemical, physical }
   }
}
\keyword{datasets}
