% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Derived_Annual_Stations}
\alias{Derived_Annual_Stations}
\title{Derived Annual station data}
\format{
A dataframe
\describe{
    \item{station}{The name of the station where data was collected}
    \item{year}{The year in which data was collected}
    \item{temperature_in_air}{Average air temperatures over annual scales at specific stations including Sable Island}
    \item{integrated_temperature_0_50}{Averages of temperature measurements, integrated between 0 and 50 metres}
    \item{integrated_salinity_0_50}{Averages of salinity measurements, integrated between 0 and 50 metres}
    \item{integrated_sigmaTheta_0_50}{Averages of density measurements, integrated between 0 and 50 metres}
    \item{stratification_0_50}{Stratification between 0 and 50 metres.}
    \item{temperature_90}{Temperature at 90 metres, or the bottom, at Prince 5 (P5).}
    \item{salinity_90}{Salinity at 90 metres, or the bottom, at Prince 5 (P5).}
    \item{sigmaTheta_90}{Potential density anomaly (sigmaTheta) at 90 metres, or the bottom, at Prince 5 (P5).}
    \item{temperature_150}{Temperature at 150 metres, or the bottom, at Halifax 2 (HL2).}
    \item{salinity_150}{Salinity at 150 metres, or the bottom, at Halifax 2 (HL2).}
    \item{sigmaTheta_150}{Potential density anomaly (sigmaTheta) at 150 metres, or the bottom, at Halifax 2 (HL2).}
    \item{sea_surface_temperature_from_moorings}{Aea surface temperatures recorded by mooring at Halifax, NS and St. Andrews, NB}
    \item{integrated_chlorophyll_0_100}{Chloropyll concentrations integrated between 0 and 100 metres.}
    \item{integrated_nitrate_0_50}{Nitrate concentrations integrated between 0 and 50 metres}
    \item{integrated_nitrate_50_150}{Nitrate concentrations integrated between 50 and 150 metres}
    \item{integrated_phosphate_0_50}{Phosphate concentrations integrated between 0 and 50 metres}
    \item{integrated_phosphate_50_150}{Phosphate concentrations integrated between 50 and 150 metres}
    \item{integrated_silicate_0_50}{Silicate concentrations integrated between 0 and 50 metres}
    \item{integrated_silicate_50_150}{Silicate concentrations integrated between 50 and 150 metres}
    \item{longitude}{Longitude (decimal degrees) of the station.}
    \item{latitude}{Latitude (decimal degrees) of the station.}

}
}
\usage{
Derived_Annual_Stations
}
\description{
Annual composites of derived data collected at fixed stations by year.
}
\details{
The data can be cited as follows:
  Casault, B., Johnson, C., Devred, E., Head, E., Cogswell, A., and
  Spry, J. 2020. Optical, Chemical, and Biological Oceanographic Conditions
  on the Scotian Shelf and in the Eastern Gulf of Maine during 2018. DFO Can.
  Sci. Advis. Sec. Res. Doc. 2020/037. v + 66 p.

  Hebert, D., Pettipas, R., and Brickman, D. 2020. Physical Oceanographic
  Conditions on the Scotian Shelf and in the Gulf of Maine during 2018. DFO
  Can. Sci. Advis. Sec. Res. Doc. 2020/036 iv + 52 p.
}
\note{
\describe{
   \item{time_scale}{Annual}
   \item{regional_scale}{Station}
   \item{category}{ biological, phytoplankton, biochemical, physical}
   }
}
\keyword{datasets}
