% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Derived_Monthly_Broadscale}
\alias{Derived_Monthly_Broadscale}
\title{Derived Monthly Broadscale data}
\format{
A dataframe
\describe{
    \item{area}{Area name where data was collected}
    \item{year}{The year in which data was collected}
    \item{month}{The month in which data was collected (numeric)}

    \item{river_flux}{A measure of the flux exiting from the St. Lawrence river into the Gulf of St. Lawrence. }
}
}
\usage{
Derived_Monthly_Broadscale
}
\description{
Derived data averaged over broad regions and months.
}
\details{
The data can be cited as follows:

  Hebert, D., Pettipas, R., and Brickman, D. 2020. Physical Oceanographic
  Conditions on the Scotian Shelf and in the Gulf of Maine during 2018. DFO
  Can. Sci. Advis. Sec. Res. Doc. 2020/036 iv + 52 p.
}
\note{
\describe{
   \item{time_scale}{Monthly}
   \item{regional_scale}{Broadscale}
   \item{category}{physical}
   }
}
\keyword{datasets}
