% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Derived_Monthly_Stations}
\alias{Derived_Monthly_Stations}
\title{Derived Monthly Station data}
\format{
A dataframe
\describe{
    \item{station}{The name of the station where data was collected}
    \item{year}{The year in which data was collected}
    \item{month}{The month in which data was collected (numeric)}


    \item{sea_surface_height}{Average of sea surface height at Yarmouth and Halifax NS}
}
}
\usage{
Derived_Monthly_Stations
}
\description{
Derived data products from fixed stations collected monthly.
}
\details{
The data can be cited as follows:
  Hebert, D., Pettipas, R., and Brickman, D. 2020. Physical Oceanographic
  Conditions on the Scotian Shelf and in the Gulf of Maine during 2018. DFO
  Can. Sci. Advis. Sec. Res. Doc. 2020/036 iv + 52 p.
}
\note{
\describe{
   \item{time_scale}{Monthly}
   \item{regional_scale}{Station}
   \item{category}{ physical}
   }
}
\keyword{datasets}
