% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Derived_Occupations_Stations}
\alias{Derived_Occupations_Stations}
\title{Derived Occupation Station data}
\format{
A dataframe with 737 observations of 14 variables
\describe{
   \item{station}{The name of the station where data was collected}
   \item{latitude}{The latitude at which data was collected}
   \item{longitude}{The longitude at which data was collected}
   \item{year}{The year in which data was collected}
   \item{month}{The month in which data was collected (numeric)}
   \item{day}{The day on which data was collected}
   \item{event_id}{A unique identifier for the sampling event}
   \item{sample_id}{A unique identifier for the sample}
   \item{depth}{The depth at which data was actually collected}
   \item{nominal_depth}{The depth at which data was planned to be collected - sometime differs slightly from actual collection depth}

   \item{mixed_layer_depth}{The depth of the well mixed surface layer at fixed stations}
   \item{density_gradient_0_50}{An index of stratification, measured as the density difference between 0 and 50 metres}
   \item{euphotic_depth}{The depth of the euphotic zone measured at fixed stations}
   \item{integrated_nitrate_0_50}{Nitrate concentrations integrated between 0 and 50 metres}
   \item{integrated_nitrate_50_150}{Nitrate concentrations integrated between 50 and 150 metres}
   \item{integrated_phosphate_0_50}{Phosphate concentrations integrated between 0 and 50 metres}
   \item{integrated_phosphate_50_150}{Phosphate concentrations integrated between 50 and 150 metres}
   \item{integrated_silicate_0_50}{Silicate concentrations integrated between 0 and 50 metres}
   \item{integrated_silicate_50_150}{Silicate concentrations integrated between 50 and 150 metres}
   \item{integrated_chlorophyll_0_100}{Chlorophyll concentrations integrated between 0 and 100 metres}
}
}
\usage{
Derived_Occupations_Stations
}
\description{
Metrics derived from data collected at fixed station locations during individual occupations.
}
\details{
The variables are as follows:

The data can be cited as follows:
  Casault, B., Johnson, C., Devred, E., Head, E., Cogswell, A., and
  Spry, J. 2020. Optical, Chemical, and Biological Oceanographic Conditions
  on the Scotian Shelf and in the Eastern Gulf of Maine during 2018. DFO Can.
  Sci. Advis. Sec. Res. Doc. 2020/037. v + 66 p.

  Hebert, D., Pettipas, R., and Brickman, D. 2020. Physical Oceanographic
  Conditions on the Scotian Shelf and in the Gulf of Maine during 2018. DFO
  Can. Sci. Advis. Sec. Res. Doc. 2020/036 iv + 52 p.
}
\note{
\describe{
   \item{time_scale}{Occupation}
   \item{regional_scale}{Station}
   \item{category}{ biological, biochemical, physical}
   }
}
\keyword{datasets}
