% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2sf.R
\name{df2sf}
\alias{df2sf}
\title{df2sf}
\usage{
df2sf(
  input = NULL,
  lat.field = "latitude",
  lon.field = "longitude",
  PID = NULL,
  ORD = NULL,
  epsg = 4326,
  type.field = NULL,
  point.IDs = NULL,
  line.IDs = NULL,
  poly.IDs = NULL,
  quiet = F
)
}
\arguments{
\item{input}{a df contain coordinates for conversion to sf.}

\item{lat.field}{the default is \code{"latitude"}. This is the name of the field holding latitude
values (in decimal degrees)}

\item{lon.field}{the default is \code{"longitude"}.  This is the name of the field holding
longitude values (in decimal degrees)}

\item{PID}{the default is \code{"PID"}.  This is a field that uniquely identifies discrete
points, lines, or polygons.}

\item{ORD}{the default is \code{NULL}.  This is not used if \code{out} is "points".  This is the
name of a field holding the order in which the provided coordinates should be joined (in lines or
polygons).  If nothing is provided, it will be assumed that the data has been provided in the
correct order.}

\item{epsg}{the default is \code{4326}.  This is the coordinate system associated with the input
positions, and it assumes WGS84 (i.e. collected via a GPS).}

\item{type.field}{the default is \code{NULL}. This is a field within the data that contains
information about whether each record is a point, line, or polygon.}

\item{point.IDs}{the default is \code{NULL}. This is a vector of values from within
\code{type.field} that can be used to recognize the point objects (e.g. "station").}

\item{line.IDs}{the default is \code{NULL}. This is a vector of values from within
\code{type.field} that can be used to recognize the line objects.}

\item{poly.IDs}{the default is \code{NULL}. This is a vector of values from within
\code{type.field} that can be used to recognize the polygon objects.}

\item{quiet}{default is \code{F}.  If 1 PID is found to have multiple positions, by default, a
message is displayed, and information about the points is shown.  If this is set to T, no message
or information is shown.}
}
\value{
an sf object
}
\description{
This function will convert a df into an sf objects.
}
\note{
This is a duplicate of what exists in https://github.com/Maritimes/Mar.utils/blob/master/R/df2sf.R.
It is copied, rather than added as a dependency to reduce the number of packages necessary.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
