% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Discrete_Annual_Broadscale}
\alias{Discrete_Annual_Broadscale}
\title{Discrete annual broadscale data}
\format{
A dataframe
\describe{
    \item{year}{The year in which data was collected}
    \item{area}{Area name where data was collected}

    \item{salinity}{Discrete salinity measurements over a range of depths}
    \item{sea_temperature}{Discrete sea temperature measurements over a range of depths}
}
}
\usage{
Discrete_Annual_Broadscale
}
\description{
Discrete data averaged over broad regions and years.
}
\details{
The variables are as follows:

The data can be cited as follows:
Hebert, D., Pettipas, R., and Brickman, D. 2020. Physical Oceanographic
  Conditions on the Scotian Shelf and in the Gulf of Maine during 2018. DFO
  Can. Sci. Advis. Sec. Res. Doc. 2020/036 iv + 52 p.
}
\note{
\describe{
   \item{time_scale}{Annual}
   \item{regional_scale}{Broadscale}
   \item{category}{ physical }
   }
}
\keyword{datasets}
