% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{Discrete_Occupations_Stations}
\alias{Discrete_Occupations_Stations}
\title{Discrete Occupation Station data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5869 rows and 14 columns.
}
\usage{
Discrete_Occupations_Stations
}
\description{
This data frame includes
_Metadata_
* station_name
* latitude
* longitude
* year
* month
* day
* event_id
* sample_id
* depth
* nominal_depth
}
\details{
_Data_
* nitrate
* silicate
* phosphate
* chlorophyll
* temperature
* salinity
* density

 Data variables are described in more detail in variable look-up table (see \code{\link{lookup_variable}})
}
\keyword{datasets}
