% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_variable.R
\name{lookup_variable}
\alias{lookup_variable}
\title{Search Variable look-up table}
\usage{
lookup_variable(variable, keyword, temporal_scale, geographic_scale, category)
}
\arguments{
\item{variable}{search by variable name}

\item{keyword}{search for a keyword within variable definitions}

\item{category}{search by data category (options are 'physical', 'biochemical', 'phenology', 'zooplankton', 'remote sensing')}

\item{time_scale}{search by temporal scale (options are 'occupations', 'weekly', 'monthly', 'annual')}

\item{regional_scale}{search by geographic scale (options are 'sections', 'stations', 'broadscale')}
}
\value{
a section of the variable look up table based on search parameters
}
\description{
Search Variable look-up table
}
