% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nitrate_inventory_timeseries_hl2}
\alias{nitrate_inventory_timeseries_hl2}
\title{Timeseries of nitrate inventory at Halifax-2 station.}
\format{A data frame with five variables: \code{year}, \code{month},
  \code{day}, \code{nitrate_0_50}, and \code{nitrate_50_150}. The nitrate
  inventory has units of mmol N/m2.}
\usage{
nitrate_inventory_timeseries_hl2
}
\description{
The nitrate inventory consists in the integrated nitrate concentration
 as obtained using trapezoidal integration over the 0-50m layer (or 50-150m)
  for individual profiles.
}
\keyword{datasets}
