% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Phytoplankton_Annual_Stations}
\alias{Phytoplankton_Annual_Stations}
\title{Phytoplankton Annual Station data}
\format{
An object of class \code{data.frame} with 42 rows and 6 columns.
}
\usage{
Phytoplankton_Annual_Stations
}
\description{
This data frame includes
_Metadata_
* station
* year
}
\details{
_Data_
* microplankton_abundance_log10
}
\keyword{datasets}
