% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_availability.R
\name{plot_availability}
\alias{plot_availability}
\title{plot_availability}
\usage{
plot_availability(areaType = NULL, areaName = NULL, parameters = NULL)
}
\arguments{
\item{areaType}{default is \code{NULL}. Indicates which type of area is of interest. The options are section, station, or area.}

\item{areaName}{default is \code{NULL}. Indicates which area is of interest. The options are dependant on the areaType, however, a error message will let the user know which options are available.}

\item{parameters}{default is \code{NULL}. Indicates which parameter is of interest. This is dependant on areaType and areaName, however, an error message will let the user know which options are available.}
}
\description{
This generates a score card of samples per month based on areaType, areaName, and parameter.
}
\examples{
\dontrun{
plot_availability(areaType="section", areaName="p5", parameter="nitrate")
}
}
\author{
Jaimie Harbin \email{jaimie.harbin@dfo-mpo.gc.ca}
}
