% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_azmpdata.R
\name{plot_azmpdata}
\alias{plot_azmpdata}
\title{Plot azmpdata dataset variable}
\usage{
plot_azmpdata(dataset, variable)
}
\arguments{
\item{dataset}{string (lowercase only)}

\item{variable}{string (lowercase only)}
}
\description{
Plot a variable from a given dataset.
}
