% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plot azmpdata dataset variable}
\usage{
plot_timeseries(ls_data)
}
\arguments{
\item{ls_data}{list}
}
\value{
ggplot2 object
}
\description{
Plot a variable from a given "timeseries" dataset.
}
\examples{
"No example"
}
