% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_physical.R
\name{read.physical}
\alias{read.physical}
\title{Read in physical data}
\usage{
read.physical(file)
}
\arguments{
\item{file}{a connection or a character string giving the name of the file to read.}
}
\description{
This function reads in all data associated with the physical oceanographic annual reporting data.
}
\details{
This function reads in all data associated with the physical oceanographic annual reporting data as
a list with metadata information and associated data.
}
\author{
Chantelle Layton
}
