% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RemoteSensing_Annual_Broadscale}
\alias{RemoteSensing_Annual_Broadscale}
\title{Remote sensing Annual broadscale data}
\format{
A dataframe
\describe{
    \item{area}{Area name where data was collected}
    \item{year}{The year in which data was collected}

    \item{surface_chlorophyll_log10}{Average surface chlorophyll as inferred from satellite (remote sensing) measurements, with a log 10 transform}
    \item{bloom_start}{Start of pythoplankton bloom, from satellite}
    \item{bloom_duration}{Length of phytoplankton bloom, from satellite}
    \item{bloom_amplitude}{Amplitude of phytoplankton bloom, from satellite}
    \item{bloom_magnitude}{Magnitude of phytoplankton bloom, from satellite}
}
}
\usage{
RemoteSensing_Annual_Broadscale
}
\description{
Remote Sensing data averaged over broad regions and years.
}
\details{
The variables are as follows:

The data can be cited as follows:
}
\note{
\describe{
   \item{time_scale}{Annual}
   \item{regional_scale}{Broadscale}
   \item{category}{ biolgocial, biochemical, satellite }
   }
}
\keyword{datasets}
