% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RemoteSensing_Weekly_Broadscale}
\alias{RemoteSensing_Weekly_Broadscale}
\title{Remote sensing weekly broadscale data}
\format{
A dataframe
\describe{
    \item{area}{Area name where data was collected}
    \item{year}{The year in which data was collected}
    \item{month}{The month in which data was collected (numeric)}
    \item{week}{The week in which data was collected (numeric)}

    \item{surface_chlorophyll}{Average surface chlorophyll as inferred from satellite (remote sensing) measurements}
}
}
\usage{
RemoteSensing_Weekly_Broadscale
}
\description{
Remote Sensing data averaged over broad regions and weeks.
}
\details{
The variables are as follows:

The data can be cited as follows:
}
\note{
\describe{
   \item{time_scale}{Weekly}
   \item{regional_scale}{Broadscale}
   \item{category}{ biological, biochemical, satellite }
   }
}
\keyword{datasets}
