% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_azmpdata.R
\name{search_azmpdata}
\alias{search_azmpdata}
\title{Search the azmpdata dataset}
\usage{
search_azmpdata(str)
}
\arguments{
\item{str}{string (lowercase only)}
}
\value{
character vector
}
\description{
Find variables in the azmpdata dataset that match a given keyword.
}
