% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{summerBottomTemperature}
\alias{summerBottomTemperature}
\title{Timeseries of averaged bottom temperature in defined NAFO zones.}
\format{
A data frame with three variables: \code{year}, \code{area_name},
\code{temperature_at_seafloor}.
}
\usage{
summerBottomTemperature
}
\description{
Timeseries of averaged bottom temperature in defined NAFO zones
along the Scotia Shelf. This uses hydrographic data from the summer groundfish
survey which generally occurs during July to early August.
}
\keyword{datasets}
