% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_check.R
\name{update_check}
\alias{update_check}
\title{updateCheck}
\usage{
update_check(gitPkg = NULL)
}
\arguments{
\item{gitPkg}{default is \code{NULL}. This is the URL to the DESCRIPTION file
on github.}
}
\description{
This function compares the package with the available version on
github, and prompts the user to update.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}, Emily Chisholm
}
\concept{general_use}
