% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_lookup.R
\name{variable_lookup}
\alias{variable_lookup}
\title{Search through RDA variable names}
\usage{
variable_lookup(keywords)
}
\arguments{
\item{keywords}{Search keywords}
}
\description{
Search through RDA variable names
}
