% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Zooplankton_Annual_Sections}
\alias{Zooplankton_Annual_Sections}
\title{Zooplankton Annual Section data}
\format{
A dataframe
\describe{
    \item{section}{Section name where data was collected}
    \item{year}{The year in which data was collected}

    \item{calanus_finmarchicus_log10}{Abundance of calanus finmarchicus with a log 10 transform }
    \item{copepods_log10}{Abundance of copepods group  with a log 10 transform }
    \item{non_copepods_log10}{Abundance of non-copepods group  with a log 10 transform }
    \item{Arctic_Calanus_species_log10}{Abundance of arctic calanus group   with a log 10 transform }
    \item{warm_offshore_copepods_log10}{Abundance of warm offshore copepods group (Clausocalanus spp., Mecynocera clausi, and Pleuromamma borealis) with a log 10 transform }
    \item{warm_shelf_copepods_log10}{Abundance of warm shelf copepods group  with a log 10 transform (the summer-fall copepods Paracalanus spp. and Centropages typicus)}
    \item{pseudocalanus_log10}{Abundance of pseudocalanus  with a log 10 transform }
    \item{zooplankton_meso_dry_weight}{Dry weight of meso zooplankton}
    \item{zooplankton_total_wet_weight}{Wet weight of all zooplankton from sampling}
}
}
\usage{
Zooplankton_Annual_Sections
}
\description{
Zooplankton data collected along AZMP transects each year.
}
\details{
The data can be cited as follows:
  Casault, B., Johnson, C., Devred, E., Head, E., Cogswell, A., and
  Spry, J. 2020. Optical, Chemical, and Biological Oceanographic Conditions
  on the Scotian Shelf and in the Eastern Gulf of Maine during 2018. DFO Can.
  Sci. Advis. Sec. Res. Doc. 2020/037. v + 66 p.
}
\note{
\describe{
   \item{time_scale}{Annual}
   \item{regional_scale}{Section}
   \item{category}{ biological, biochemical, zooplankton }
   }
}
\keyword{datasets}
