% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Zooplankton_Seasonal_Broadscale}
\alias{Zooplankton_Seasonal_Broadscale}

\title{Zooplankton seasonal broadscale data}
\format{
A dataframe containing 40 observations of 6 variables:
\describe{
   \item{region}{The region in which data was collected}
   \item{year}{The year in which data was collected}
   \item{season}{The season in which data was collected (Northern Hemisphere)}


   \item{Calanus_finmarchicus_log10}{Abundance of Calanus finmarchicus with a log10 transform}
   \item{zooplankton_meso_dry_weight}{Dry weight of meso zooplankton}
   \item{zooplankton_total_wet_weight}{Wet weight of all zooplankton sampled}
}

 @note
\describe{
   \item{time_scale}{Seasonal}
   \item{regional_scale}{Broadscale}
   \item{category}{Zooplankton, biological, biochemical}
   }
}

\usage{
Zooplankton_Seasonal_Broadscale
}
\description{
A dataframe containing seasonal broadscale zooplankton data.
The variables are as follows:
}
\details{
The data can be cited as follows:
  Casault, B., Johnson, C., Devred, E., Head, E., Cogswell, A., and
  Spry, J. 2020. Optical, Chemical, and Biological Oceanographic Conditions
  on the Scotian Shelf and in the Eastern Gulf of Maine during 2018. DFO Can.
  Sci. Advis. Sec. Res. Doc. 2020/037. v + 66 p.
}
\keyword{datasets}
