% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_indexer.R
\name{area_indexer}
\alias{area_indexer}
\title{area_indexer}
\usage{
area_indexer(
  years = NULL,
  areanames = NULL,
  areaTypes = NULL,
  datafiles = NULL,
  months = NULL,
  doMonths = F,
  doParameters = F,
  parameters = NULL,
  fuzzyParameters = TRUE,
  qcMode = F
)
}
\arguments{
\item{years}{default is \code{NULL}.  If you want to restrict the available data by one or more
years, a vector of desired years can be provided (e.g. \code{area_indexer(years=c(2017,2018))})}

\item{areanames}{default is \code{NULL}.  If you want to restrict the available data by one or
more specific named areas, a vector of them can be provided (e.g. \code{area_indexer(areanames=c("HL","HL2"))})}

\item{areaTypes}{default is \code{NULL}.  If you want to restrict the available data by one or more
area "types", a vector of desired types can be provided.  Valid values for areaTypes are:
"area", "section", and "station".  (e.g. \code{area_indexer(areaTypes=c("section"))})}

\item{datafiles}{default is \code{NULL}.  If you want to restrict the available data to just that
contained by one or more particular azmpdata data files, a vector of the files to check can be
provided.  .  (e.g. \code{area_indexer(datafiles=c("Ice_Annual_Broadscale"))})  A complete list
of the available files can be seen by checking \code{data(package="azmpdata")}}

\item{months}{default is \code{NULL}.  If you want to restrict the available data by one or more
months, a vector of desired months can be provided (e.g. \code{area_indexer(months=c(1,2,3,4))})}

\item{doMonths}{default is \code{F}. If this is set to \code{TRUE}, the results will include
information about what month the data was collected (when available).}

\item{doParameters}{default is \code{F}.  Identifying all of the parameters that were collected
at each combination of area/site/year is a bit more intensive than leaving them out.  To force
this function to do this, set this parameter to \code{doParameters=T}.  If one or more parameters
are provided, the function will override this default value and set \code{doParameters=F}}

\item{parameters}{default is \code{NULL}.  Many parameters exist - any named parameter found in
any data file should work.  (e.g. \code{area_indexer(parameters=c("Arctic_Calanus_species",
"integrated_phosphate_50"))})}

\item{fuzzyParameters}{default is \code{T}.  By default, any discovered parameters that match
values within \code{parameters} will be returned.  For example, \code{parameter="nitrate"} will
return fields such as "integrated_nitrate_0_50", "integrated_nitrate_50_150", and "nitrate".  If
you want exact matches only, set \code{fuzzyParameters = F}.}

\item{qcMode}{default is \code{F}. Information about unexpected values will be shown.}
}
\value{
a data.frame
}
\description{
This function assembles a dataframe consisting of the years and areas where the azmpdata
packages\'s data has been collected, as well as the associated file(s) where the data can be found.
}
\note{
Note that each additional filter that gets sent will reduce the number of results returned.
For example, if \code{doMonths = TRUE} and \code{years = 2010}, only those results from 2010 that
also have monthly data will be returned.
}
\examples{
\dontrun{
allAreas <- area_indexer()

allAreas_w_Parameters <- area_indexer(doParameters =T)

areaAreas_2018_2020 <- area_indexer(years=c(2018,2019,2020))

some_HLData <- area_indexer(areanames=c("HL","HL2"))

sections_2017 <- area_indexer(areaTypes=c("section"), year = 2017)

specificParameters_2000s <- area_indexer(parameters=c("Arctic_Calanus_species",
                                         "integrated_phosphate_0_50"), year=c(2000:2009))
februaryParameters <-area_indexer(doMonths = T, months = 2, doParameters = T)
}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
