% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azmpdata.R
\docType{package}
\name{azmpdata}
\alias{azmpdata}
\alias{azmpdata-package}
\title{azmpdata: A package providing data products derived from data collected as part of
The Atlantic Zone Monitoring Program (AZMP).}
\description{
The azmp package provides three categories of data products:
physical, biological and chemical variables.  For a complete list of
variables, use library(help = "azmpdata").
}
\section{Physical variables}{

The physical variables include temperature, salinity, ...
}

\section{Biological variables}{

The biological variables include chlorophyll, phyto- and zooplankton
abundance, and zooplankton biomass.
}

\section{Chemical variables}{

The chemical variables in main nutrients (nitrate, silicate and phosphate).
}

