% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azmpmap.R
\name{azmpmap}
\alias{azmpmap}
\title{azmpmap}
\usage{
azmpmap(quiet = T)
}
\arguments{
\item{quiet}{default is \code{T} This indicates whether or not you want to see messages related
to the transformation of data from the ftp site to spatial objects.}
}
\description{
This generates a leaflet plot of the geospatial files stored at
ftp://ftp.dfo-mpo.gc.ca/AZMP_Maritimes/azmpdata/data/lookup/polygons/.
}
\examples{
\dontrun{
map <- azmpmap()
}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
