% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Derived_Annual_Broadscale}
\alias{Derived_Annual_Broadscale}
\title{Derived annual broadscale data}
\format{
A dataframe
\describe{
    \item{area}{Area name where data was collected}
    \item{year}{The year in which data was collected}

    \item{sea_temperature_at_sea_floor_july}{Temperature at sea floor during the month of July, averaged over NAFO areas 4X, 4W and 4V.}
    \item{sea_surface_temperature}{Temperature at sea surface averaged over NAFO regions 4V, 4X and 4W as well as regional areas (Bay of Fundy + Gulf of Maine)}
    \item{north_atlantic_oscillation}{North Atlantic Oscillation Index (NAO), calculated using sea level pressure difference between subtropical high and subpolar low}
    \item{atlantic_multidecadal_oscillation}{Atlantic Multidecadal Oscillation Index (AMO), based on anomalies of sea surface temperatures in North Atlantic basin. }
    \item{cold_intermediate_layer_volume}{A measure of the volume of the cold intermediate layer throughout the Gulf of St. Lawrence}
    \item{sea_ice_volume}{Average ice volume for January - March in the Gulf of St. Lawrence, can be reasonably used as a proxy to represent sea ice volume on the Scotian Shelf. }
    \item{density_gradient_0_50}{An index of stratification, measured as the density difference between 0 and 50 metres}
    \item{temperature_at_sea_floor}{Averaged water temperatures at sea floor, over NAFO areas 4X, 4V and 4W}
    \item{minimum_temperature_in_cold_intermediate_layer}{Minimum temperature in the cold intermediate layer within the Gulf of St. Lawrence }
    \item{salinity}{Discrete salinity measurements over a range of depths}
    \item{sea_temperature}{Discrete sea temperature measurements over a range of depths}
    \item{sea_surface_temperature_from_satellite}{Averages of sea surface temperatures infered from remote sensing (satellite) measurements}
    \item{sea_surface_temperature_warming}{Derived metric calculating the warming of sea surface temperature averages over the Scotian Shelf}
    \item{final_day_of_sea_ice}{The final day of sea ice observed in the Gulf of St. Lawrence. Note that this metric can be used as a reasonable proxy for the Scotian Shelf}
    \item{sea_ice_area}{Average  of sea ice area, seaward of Cabot Strait}
    \item{surface_chlorphyll}{Average surface chlorophyll as inferred from satellite (remote sensing) measurements}
}
}
\usage{
Derived_Annual_Broadscale
}
\description{
Derived data averaged over broad regions and years.
}
\details{
The data can be cited as follows:
  Casault, B., Johnson, C., Devred, E., Head, E., Cogswell, A., and
  Spry, J. 2020. Optical, Chemical, and Biological Oceanographic Conditions
  on the Scotian Shelf and in the Eastern Gulf of Maine during 2018. DFO Can.
  Sci. Advis. Sec. Res. Doc. 2020/037. v + 66 p.

  Hebert, D., Pettipas, R., and Brickman, D. 2020. Physical Oceanographic
  Conditions on the Scotian Shelf and in the Gulf of Maine during 2018. DFO
  Can. Sci. Advis. Sec. Res. Doc. 2020/036 iv + 52 p.
}
\note{
\describe{
   \item{time_scale}{Annual}
   \item{regional_scale}{Broadscale}
   \item{category}{ biological, biochemical, physical }
   }
}
\keyword{datasets}
