% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Derived_Annual_Sections}
\alias{Derived_Annual_Sections}
\title{Derived Annual Section data}
\format{An object of class \code{data.frame} with 82 rows and 9 columns.}
\usage{
Derived_Annual_Sections
}
\description{
_Metadata_
* section
* year
}
\details{
_Data_
* density
* temperature
* salinity
* integrated_nitrate_0_50
* integrated_nitrate_50_150
* integrated_phosphate_0_50
* integrated_phosphate_50_150
* integrated_silicate_0_50
* integrated_silicate_50_150
* integrated_chlorphyll_0_100

Data variables are described in more detail in variable look-up table (see \code{\link{lookup_variable}})
}
\keyword{datasets}
