% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{Derived_Annual_Stations}
\alias{Derived_Annual_Stations}
\title{Derived Occupation Monthly data}
\format{An object of class \code{data.frame} with 1004 rows and 12 columns.}
\usage{
Derived_Annual_Stations
}
\description{
This data frame includes
_Metadata_
* station_name
* latitude
* longitude
* year
* month
}
\details{
_Data_
* sea_surface_height

<<<<<<< HEAD
Data variables are described in more detail in variable look-up table (see \code{\link{lookup_variable()}})
=======
 Data variables are described in more detail in variable look-up table (see \code{\link{lookup_variable}})
>>>>>>> d1a556953fcf1189b8625cb6a8034e8b0fdb0fdd
Derived Occupation Annual data

This data frame includes
_Metadata_
* station_name
* latitude
* longitude
* year

_Data_
* temperature_in_air
* denisty
* sea_temperature
* salinity
* integrated_sea_temperature_0_50
* integrated_salinity_0_50
* integrated_density_0_50
* temperature_0
* temperature_90
* sea_surface_temperature_from_moorings
* sea_surface_height
* integrated_nitrate_0_50
* integrated_nitrate_50_150
* integrated_phosphate_0_50
* integrated_phosphate_50_150

<<<<<<< HEAD
Data variables are described in more detail in variable look-up table (see \code{\link{lookup_variable()}})
=======

Data variables are described in more detail in variable look-up table (see \code{\link{lookup_variable}})
>>>>>>> d1a556953fcf1189b8625cb6a8034e8b0fdb0fdd
}
\keyword{datasets}
