% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Derived_Annual_Stations}
\alias{Derived_Annual_Stations}
\title{Derived Annual station data}
\format{
A dataframe
\describe{
    \item{station}{The name of the station where data was collected}
    \item{year}{The year in which data was collected}

    \item{temperature_in_air}{Average air temperatures over annual scales at specific stations including Sable Island}
    \item{denisty}{Discrete density measurements over a range of depths}
    \item{sea_temperature}{Discrete sea temperature measurements over a range of depths}
    \item{salinity}{Discrete salinity measurements over a range of depths}
    \item{integrated_sea_temperature_0_50}{Averages of temperature measurements, integrated between 0 and 50 metres}
    \item{integrated_salinity_0_50}{Averages of salinity measurements, integrated between 0 and 50 metres}
    \item{integrated_density_0_50}{Averages of density measurements, integrated between 0 and 50 metres}
    \item{temperature_0}{Averages for temperature at 0 metres at Prince 5 (P5) and Halifax 2 (HL2) stations}
    \item{temperature_90}{Averages for temperature at 90 metres at Prince 5 (P5) and Halifax 2 (HL2) stations}
    \item{sea_surface_temperature_from_moorings}{Averages of sea surface temperatures recorded by mooring at Halifax, NS and St. Andrews, NB}
    \item{sea_surface_height}{Average of sea surface height at Yarmouth and Halifax NS}
    \item{integrated_nitrate_0_50}{Nitrate concentrations integrated between 0 and 50 metres}
    \item{integrated_nitrate_50_150}{Nitrate concentrations integrated between 50 and 150 metres}
    \item{integrated_phosphate_0_50}{Phosphate concentrations integrated between 0 and 50 metres}
    \item{integrated_phosphate_50_150}{Phosphate concentrations integrated between 50 and 150 metres}
}
}
\usage{
Derived_Annual_Stations
}
\description{
Derived data collected at fixed stations by year.
}
\details{
The data can be cited as follows:
  Casault, B., Johnson, C., Devred, E., Head, E., Cogswell, A., and
  Spry, J. 2020. Optical, Chemical, and Biological Oceanographic Conditions
  on the Scotian Shelf and in the Eastern Gulf of Maine during 2018. DFO Can.
  Sci. Advis. Sec. Res. Doc. 2020/037. v + 66 p.

  Hebert, D., Pettipas, R., and Brickman, D. 2020. Physical Oceanographic
  Conditions on the Scotian Shelf and in the Gulf of Maine during 2018. DFO
  Can. Sci. Advis. Sec. Res. Doc. 2020/036 iv + 52 p.
}
\note{
\describe{
   \item{time_scale}{Annual}
   \item{regional_scale}{Station}
   \item{category}{ biological, phytoplankton, biochemical, physical}
   }
}
\keyword{datasets}
