% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{Derived_Occupations_Sections}
\alias{Derived_Occupations_Sections}
\title{Derived Occupation Section data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1103 rows and 15 columns.
}
\usage{
Derived_Occupations_Sections
}
\description{
This data frame includes
_Metadata_
* section_name
* station_name
* latitude
* longitude
* year
* month
* day
* event_id
* cruise_id
* sample_id
* depth
* nominal_depth
* odf_filename
}
\details{
_Data_
* integrated_nitrate_0_50
* integrated_nitrate_50_150
* integrated_phosphate_0_50
* integrated_phosphate_50_150
* integrated_silicate_0_50
* integrated_silicate_50_150
* integrated_chlorophyll_0_100

 Data variables are described in more detail in variable look-up table (see \code{\link{lookup_variable()}})
}
\keyword{datasets}
