% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{Derived_Occupations_Stations}
\alias{Derived_Occupations_Stations}
\title{Derived Occupation Station data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10992 rows and 11 columns.}
\usage{
Derived_Occupations_Stations
}
\description{
Metrics derived from data collected at fixed station locations during individual occupations.
}
\details{
This data frame includes
_Metadata_
* station_name
* latitude
* longitude
* year
* month
* day
* event_id
* sample_id
* depth
* nominal_depth

_Data_
* mixed_layer_depth
* density_gradient_0_50
* euphotic_depth
* integrated_nitrate_0_50
* integrated_nitrate_50_150
* integrated_phosphate_0_50
* integrated_phosphate_50_150
* integrated_silicate_0_50
* integrated_silicate_50_150
* integrated_chlorophyll_0_100

Data variables are described in more detail in variable look-up table (see \code{\link{lookup_variable}})
}
\keyword{datasets}
