% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Discrete_Annual_Broadscale}
\alias{Discrete_Annual_Broadscale}
\title{Discrete annual broadscale data}
\format{
An object of class \code{data.frame} with 72 rows and 4 columns.
}
\usage{
Discrete_Annual_Broadscale
}
\description{
_Metadata_
* year
* area
}
\details{
_Data_
* salinity
* sea_temperature
Data variables are described in more detail in variable look-up table (see \code{\link{lookup_variable}})
}
\keyword{datasets}
