% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Discrete_Occupations_Sections}
\alias{Discrete_Occupations_Sections}
\title{Discrete Occupation Section data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12055 rows and 15 columns.
}
\usage{
Discrete_Occupations_Sections
}
\description{
This data frame includes
_Metadata_
* section
* station
* latitude
* longitude
* year
* month
* day
* event_id
* cruise_id
* sample_id
* depth
* nominal_depth
* odf_filename
}
\details{
_Data_
* nitrate
* silicate
* phosphate
* chlorophyll
* temperature
* salinity
* density

Data variables are described in more detail in variable look-up table (see \code{\link{lookup_variable}})
}
\keyword{datasets}
