% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Phytoplankton_Annual_Stations}
\alias{Phytoplankton_Annual_Stations}
\title{Phytoplankton Annual Station data}
\format{An object of class \code{data.frame} with 42 rows and 6 columns.}
\usage{
Phytoplankton_Annual_Stations
}
\description{
This data frame includes
_Metadata_
* station
* latitude
* longitude
* year
}
\details{
_Data_
* microplankton_abundance_log10

Data variables are described in more detail in variable look-up table (see \code{\link{lookup_variable}})
}
\keyword{datasets}
