% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{Phytoplankton_Occupations_Stations}
\alias{Phytoplankton_Occupations_Stations}
\title{Phytoplankton Occupation Station data}
\format{
An object of class \code{data.frame} with 659 rows and 13 columns.
}
\usage{
Phytoplankton_Occupations_Stations
}
\description{
This data frame includes
_Metadata_
* station_name
* latitude
* longitude
* year
* month
* day
* event_id
* sample_id
* depth
* nominal_depth
}
\details{
_Data_
* microplankton_abundance

 Data variables are described in more detail in variable look-up table (see \code{\link{lookup_variable}})
}
\keyword{datasets}
