% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Phytoplankton_Occupations_Stations}
\alias{Phytoplankton_Occupations_Stations}
\title{Phytoplankton Occupation Station data}
\format{
A dataframe
\describe{
    \item{station}{The name of the station where data was collected}
    \item{latitude}{The latitude at which data was collected}
    \item{longitude}{The longitude at which data was collected}
    \item{year}{The year in which data was collected}
    \item{month}{The month in which data was collected (numeric)}
    \item{day}{The day on which data was collected}
    \item{event_id}{A unique identifier for the sampling event}
    \item{sample_id}{A unique identifier for the sample}
    \item{depth}{The depth at which data was actually collected}

    \item{microplankton_abundance}{Abundance of microplankton at fixed stations (phytoplankton and protists)}
}
}
\usage{
Phytoplankton_Occupations_Stations
}
\description{
Phytoplankton data collected at fixed stations during individual occupations.
}
\details{
The variables are as follows:

The data can be cited as follows:
  Casault, B., Johnson, C., Devred, E., Head, E., Cogswell, A., and
  Spry, J. 2020. Optical, Chemical, and Biological Oceanographic Conditions
  on the Scotian Shelf and in the Eastern Gulf of Maine during 2018. DFO Can.
  Sci. Advis. Sec. Res. Doc. 2020/037. v + 66 p.
}
\note{
\describe{
   \item{time_scale}{Occupation}
   \item{regional_scale}{Station}
   \item{category}{ biological, phytoplankton}
   }
}
\keyword{datasets}
