% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_annual_means.R
\name{plot_annual_means}
\alias{plot_annual_means}
\title{Plot azmpdata dataset variable}
\usage{
plot_annual_means(ls_data)
}
\arguments{
\item{ls_data}{list}
}
\value{
ggplot2 object
}
\description{
Plot a variable from a given "annual means" dataset.
}
\examples{
"No example"
}
