% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plot azmpdata dataset variable}
\usage{
plot_timeseries(df_data, variable)
}
\arguments{
\item{df_data}{dataframe loaded from azmpdata}

\item{variable}{variable from dataframe to be plotted}
}
\value{
ggplot2 object
}
\description{
Plot a variable from a given "timeseries" dataset.
}
\examples{
"No example"
}
