% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RemoteSensing_Annual_Broadscale}
\alias{RemoteSensing_Annual_Broadscale}
\title{Remote sensing Annual broadscale data}
\format{
An object of class \code{data.frame} with 108 rows and 7 columns.
}
\usage{
RemoteSensing_Annual_Broadscale
}
\description{
_Metadata_
* area
* year
}
\details{
_Data_
* surface_chlorophyll_log10
* bloom_start
* bloom_duration
* bloom_amplitude
* bloom_magnitude

Data variables are described in more detail in variable look-up table (see \code{\link{lookup_variable}})
}
\keyword{datasets}
