% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RemoteSensing_Weekly_Broadscale}
\alias{RemoteSensing_Weekly_Broadscale}
\title{Remote sensing weekly broadscale data}
\format{
A dataframe
\describe{
    \item{area}{Area name where data was collected}
    \item{year}{The year in which data was collected}
    \item{month}{The month in which data was collected (numeric)}
    \item{week}{The week in which data was collected (numeric)}

    \item{surface_chlorophyll}{Average surface chlorophyll as inferred from satellite (remote sensing) measurements}
}
}
\usage{
RemoteSensing_Weekly_Broadscale
}
\description{
Remote Sensing data averaged over broad regions and weeks.
}
\details{
The data can be cited as follows:
Casault, B., Johnson, C., Devred, E., Head, E., Cogswell, A., and
  Spry, J. 2020. Optical, Chemical, and Biological Oceanographic Conditions
  on the Scotian Shelf and in the Eastern Gulf of Maine during 2018. DFO Can.
  Sci. Advis. Sec. Res. Doc. 2020/037. v + 66 p.
}
\note{
\describe{
   \item{time_scale}{Weekly}
   \item{regional_scale}{Broadscale}
   \item{category}{ biological, biochemical, satellite }
   }
}
\keyword{datasets}
