% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RemoteSensing_Weekly_Broadscale}
\alias{RemoteSensing_Weekly_Broadscale}
\title{Remote Sensing Annual Broadscale data}
\format{
An object of class \code{data.frame} with 5016 rows and 5 columns.
}
\usage{
RemoteSensing_Weekly_Broadscale
}
\description{
_Metadata_
* area
* year
}
\details{
_Data_
* surface_chlorophyll_log10
* bloom_start
* bloom_duration
* bloom_amplitude
* bloom_magnitude

Data variables are described in more detail in variable look-up table (see \code{\link{lookup_variable}})
Remote sesning weekly broadscale data

_Metadata_
* area
* year
* month
* week

_Data_
* surface_chlorophyll
}
\keyword{datasets}
