% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_lookup.R
\name{variable_lookup}
\alias{variable_lookup}
\title{Search through azmpdata}
\usage{
variable_lookup(keywords, search_help = FALSE, lib.loc = NULL)
}
\arguments{
\item{keywords}{Search keywords (if using multiple, create a vector of character strings using `c()`)}

\item{search_help}{a logical value determining whether or not help documentation text should also be searched}

\item{lib.loc}{(optional) passed to find.package to find help documentation
through which to search - a character vector describing the location of R
library trees to search through, or NULL. The default value of NULL
corresponds to checking the loaded namespace, then all libraries
currently known in .libPaths().}
}
\description{
Search through azmpdata datasets by keywords. Default searches through all
dataset variable names, but options (`help_search`) can be used to search
through all `azmpdata` help documentation text including variable definitions
and metadata.
}
\examples{
\dontrun{
res <- variable_lookup('nitrate')
res2 <- variable_lookup("nitrate", search_help = T)
}

}
