% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{Zooplankton_Occupations_Broadscale}
\alias{Zooplankton_Occupations_Broadscale}
\title{Zooplankton Occupations Broadscale}
\format{
An object of class \code{data.frame} with 1177 rows and 10 columns.
}
\usage{
Zooplankton_Occupations_Broadscale
}
\description{
_Metadata_
* area
* year
* month
* day
}
\details{
_Data_
* calanus_finmarchicus_abundance
* zooplankton_meso_dry_weight
* zooplankton_total_wet_weight

Data variables are described in more detail in variable look-up table (see \code{\link{lookup_variable}})
}
\keyword{datasets}
