% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Zooplankton_Occupations_Broadscale}
\alias{Zooplankton_Occupations_Broadscale}

\title{Zooplankton Occupation Broadscale data}
\format{
A dataframe containing 1177 obesrvations of 10 variables
\describe{
   \item{latitude}{The latitude at which data was collected}
   \item{longitude}{The longitude at which data was collected}
   \item{year}{The year in which data was collected}
   \item{month}{The month in which data was collected (numeric)}
   \item{day}{The day on which data was collected}
   \item{season}{The season in which data was collected (Northern Hemisphere)}
   \item{sample_id}{A unique identifier for each sample}

   \item{Calanus_finmarchicus_abundance}{Abundance of Calanus finmarchicus}
   \item{zooplankton_meso_dry_weight}{Dry weight of meso zooplankton}
   \item{zooplankton_total_wet_weight}{Wet weight of all zooplankton sampled}
}

 @note
\describe{
   \item{time_scale}{Occupation}
   \item{regional_scale}{Broadscale}
   \item{category}{Zooplankton, biological, biochemical}
   }
}
\usage{
Zooplankton_Occupations_Broadscale
}
\description{
A dataframe containing seasonal broadscale zooplankton data.
The variables are as follows:
}
\details{
The data can be cited as follows:
  Casault, B., Johnson, C., Devred, E., Head, E., Cogswell, A., and
  Spry, J. 2020. Optical, Chemical, and Biological Oceanographic Conditions
  on the Scotian Shelf and in the Eastern Gulf of Maine during 2018. DFO Can.
  Sci. Advis. Sec. Res. Doc. 2020/037. v + 66 p.
}
\keyword{datasets}
