% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Zooplankton_Occupations_Sections}
\alias{Zooplankton_Occupations_Sections}
\title{Zooplankton Occupation Section data}
\format{
A dataframe
\describe{
    \item{section}{Section name where data was collected}
    \item{station}{The name of the station where data was collected}
    \item{latitude}{The latitude at which data was collected}
    \item{longitude}{The longitude at which data was collected}
    \item{year}{The year in which data was collected}
    \item{season}{The season in which data was collected (Northern Hemisphere)}
    \item{month}{The month in which data was collected (numeric)}
    \item{day}{The day on which data was collected}
    \item{event_id}{A unique identifier for the sampling event}
    \item{cruise_id}{A unique identifier for the mission/cruise}
    \item{sample_id}{A unique identifier for the sample}
    \item{depth}{The depth at which data was actually collected}

    \item{calanus_finmarchicus_abundance}{Abundance of calanus finmarchicus }
    \item{pseudocalanus_abundance}{Abundance of pseudocalanus  }
    \item{copepods}{Abundance of copepods group  with a log 10 transform }
    \item{non_copepods}{Abundance of non-copepods group  }
    \item{Arctic_Calanus_species}{Abundance of arctic calanus group  }
    \item{warm_offshore_copepods}{Abundance of warm offshore copepods group (Clausocalanus spp., Mecynocera clausi, and Pleuromamma borealis)}
    \item{warm_shelf_copepods}{Abundance of warm shelf copepods group  (the summer-fall copepods Paracalanus spp. and Centropages typicus)}
    \item{zooplankton_meso_dry_weight}{Dry weight of meso zooplankton}
    \item{zooplankton_meso_wet_weight}{Wet weight of meso zooplankton}
    \item{zooplankton_macro_wet_weight}{Wet weight of macro zooplankton}
    \item{zooplankton_total_wet_weight}{Wet weight of all zooplankton from sampling}
}
}
\usage{
Zooplankton_Occupations_Sections
}
\description{
Zooplankton data collected along AZMP transects during individual occupations.
}
\details{
The data can be cited as follows:
  Casault, B., Johnson, C., Devred, E., Head, E., Cogswell, A., and
  Spry, J. 2020. Optical, Chemical, and Biological Oceanographic Conditions
  on the Scotian Shelf and in the Eastern Gulf of Maine during 2018. DFO Can.
  Sci. Advis. Sec. Res. Doc. 2020/037. v + 66 p.
}
\note{
\describe{
   \item{time_scale}{Occupation}
   \item{regional_scale}{Section}
   \item{category}{ biological, zooplankton }
   }
}
\keyword{datasets}
