% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{Zooplankton_Seasonal_Broadscale}
\alias{Zooplankton_Seasonal_Broadscale}
\title{Zooplankton Seasonal Broadscale}
\format{
An object of class \code{data.frame} with 40 rows and 6 columns.
}
\usage{
Zooplankton_Seasonal_Broadscale
}
\description{
_Metadata_
* area
* year
* season
}
\details{
_Data_
* calanus_finmarchicus_log10
* zooplankton_meso_dry_weight
* zooplankton_total_wet_Weight

Data variables are described in more detail in variable look-up table (see \code{\link{lookup_variable}})
}
\keyword{datasets}
